refactor: Refactor stories paths and update dependencies
