+++ src/stories/colors.stories.tsx
